Strip down plutip to plutip-core
This change leaves plutip as a tool for spawning local clusters,
removing all the testing functionalities for testing contracts.

 - remove functionality to run test contracts
 - sync deps to cardano-node-1.35.4
 - remove BPI and plutus-apps from deps
 - remove setup for BPI from cluster setup
 - refactor cluster setup
 - new: withCluster (previously withPlutusInterface), withFundedCluster
 - change module structure
