--- contract-execution/README.md
@@ -1,3 +1,0 @@
-# Contract execution example
-
-An example of how to start private network from Haskell and execute arbitrary contract.
