--- local-cluster-info.json
@@ -1,1 +1,0 @@
-{"ciNodeSocket":"/tmp/nix-shell.uN2SS9/test-cluster1065449/pool-1/node.socket","ciWallets":[["466ffa83718224951ebfaf9715e34404f354a4e7ce491ffcd08dbac0","\"addr1v9rxl75rwxpzf9g7h7hew90rgsz0x49yul8yj8lu6zxm4sqtss4rg\""]]}
