+++ local-cluster/Main.hs
@@ -1,6 +1,7 @@
 {-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE ImportQualifiedPost #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE NumericUnderscores #-}
@@ -8,44 +9,32 @@
 module Main (main) where
 
 import Cardano.Launcher.Node (CardanoNodeConn, nodeSocketFile)
-import Cardano.Ledger.Slot (EpochSize (EpochSize))
-import Control.Applicative (optional, (<**>), (<|>))
-import Control.Monad (forM_, replicateM, void)
+import Control.Applicative (optional, (<**>))
+import Control.Monad (forM_)
 import Control.Monad.IO.Class (liftIO)
-import Control.Monad.Reader (MonadReader (ask), ReaderT (ReaderT), lift)
 import Data.Aeson (FromJSON, ToJSON, encodeFile)
 import Data.Default (def)
+import Data.Foldable (for_)
 import Data.Time (NominalDiffTime)
-import GHC.Conc (TVar, threadDelay)
+import GHC.Conc (threadDelay)
 import GHC.Generics (Generic)
 import GHC.Natural (Natural)
 import GHC.Word (Word64)
-import Numeric.Positive (Positive)
 import Options.Applicative (Parser, helper, info)
 import Options.Applicative qualified as Options
-import System.Posix (Handler (CatchOnce), installHandler, sigINT)
-import Test.Plutip.Config (
-  ChainIndexMode (CustomPort, DefaultPort, NotNeeded),
-  PlutipConfig (chainIndexMode, clusterWorkingDir, extraConfig),
+import Plutip.Cluster (
+  dieOnError,
+  withFundedCluster,
+ )
+import Plutip.Config (
+  EpochSize (EpochSize),
+  ExtraConfig (ExtraConfig),
+  PlutipConfig (clusterWorkingDir, extraConfig),
   WorkingDirectory (Fixed, Temporary),
  )
-import Test.Plutip.Internal.BotPlutusInterface.Wallet (
-  BpiWallet,
-  addSomeWalletDir,
-  cardanoMainnetAddress,
-  mkMainnetAddress,
-  walletPkh,
- )
-import Test.Plutip.Internal.Cluster.Extra.Types (
-  ExtraConfig (ExtraConfig),
- )
-import Test.Plutip.Internal.LocalCluster (
-  ClusterStatus,
-  startCluster,
-  stopCluster,
- )
-import Test.Plutip.Internal.Types (nodeSocket)
-import Test.Plutip.Tools.CardanoApi (awaitAddressFunded)
+import Plutip.DistributeFunds (Lovelace)
+import Plutip.Keys (KeyPair, mainnetAddress, saveKeyPair, showPkh)
+import Plutip.Types (nodeSocket)
 
 main :: IO ()
 main = do
@@ -53,81 +42,61 @@ main = do
   case totalAmount config of
     Left e -> error e
     Right amt -> do
-      let ClusterConfig {numWallets, dirWallets, numUtxos, workDir, slotLength, epochSize, cIndexMode} = config
+      let ClusterConfig {numWallets, dirWallets, numUtxos, workDir, slotLength, epochSize} = config
           workingDir = maybe Temporary (`Fixed` False) workDir
 
           extraConf = ExtraConfig slotLength epochSize
-          plutipConfig = def {clusterWorkingDir = workingDir, extraConfig = extraConf, chainIndexMode = cIndexMode}
+          plutipConfig = def {clusterWorkingDir = workingDir, extraConfig = extraConf}
 
       putStrLn "Starting cluster..."
-      (st, _) <- startCluster plutipConfig $ do
-        ws <- initWallets numWallets numUtxos amt dirWallets
-        liftIO $ putStrLn "Waiting for wallets to be funded..."
-        awaitFunds ws slotLength
+      withFundedCluster plutipConfig (replicate numWallets $ replicate numUtxos amt) $ \cenv keys -> do
+        -- Save keys to requested directory
+        forM_ dirWallets $ \dir -> for_ keys (fmap dieOnError . saveKeyPair dir)
 
+        -- print info
         separate
-        liftIO $ forM_ (zip ws [(1 :: Int) ..]) printWallet
-        printNodeRelatedInfo
+        liftIO $ forM_ (zip keys [(1 :: Int) ..]) printWallet
+        printNodeRelatedInfo cenv
         separate
 
+        -- Dump cluster info to local-cluster.info
         forM_ (dumpInfo config) $ \dInfo -> do
-          cEnv <- ask
-          lift $
-            dumpClusterInfo
-              dInfo
-              (nodeSocket cEnv)
-              ws
+          dumpClusterInfo
+            dInfo
+            (nodeSocket cenv)
+            keys
 
-      void $ installHandler sigINT (termHandler st) Nothing
-      putStrLn "Cluster is running. Ctrl-C to stop."
-      loopThreadDelay
+        putStrLn "Cluster is running. Ctrl-C to stop."
+        loopThreadDelay
   where
     loopThreadDelay = threadDelay 100000000 >> loopThreadDelay
 
-    printNodeRelatedInfo = ReaderT $ \cEnv -> do
+    printNodeRelatedInfo = \cEnv -> do
       putStrLn $ "Node socket: " <> show (nodeSocket cEnv)
 
     separate = liftIO $ putStrLn "\n------------\n"
 
-    totalAmount :: ClusterConfig -> Either String Positive
+    totalAmount :: ClusterConfig -> Either String Lovelace
     totalAmount cwc =
       case toAda (adaAmount cwc) + lvlAmount cwc of
         0 -> Left "One of --ada or --lovelace arguments should not be 0"
         amt -> Right $ fromInteger . toInteger $ amt
 
-    initWallets numWallets numUtxos amt dirWallets = do
-      let collateralAmount = 10_000_000
-      replicateM (max 0 numWallets) $
-        addSomeWalletDir
-          (collateralAmount : replicate numUtxos amt)
-          dirWallets
-
-    dumpClusterInfo :: FilePath -> CardanoNodeConn -> [BpiWallet] -> IO ()
+    dumpClusterInfo :: FilePath -> CardanoNodeConn -> [KeyPair] -> IO ()
     dumpClusterInfo fp nodeConn ws = do
       encodeFile
         fp
         ( ClusterInfo
-            { ciWallets = [(show . walletPkh $ w, show . mkMainnetAddress $ w) | w <- ws]
+            { ciWallets = [(showPkh w, show . mainnetAddress $ w) | w <- ws]
             , ciNodeSocket = nodeSocketFile nodeConn
             }
         )
 
     printWallet (w, n) = do
-      putStrLn $ "Wallet " ++ show n ++ " PKH: " ++ show (walletPkh w)
-      putStrLn $ "Wallet " ++ show n ++ " mainnet address: " ++ show (mkMainnetAddress w)
+      putStrLn $ "Wallet " ++ show n ++ " PKH: " ++ showPkh w
+      putStrLn $ "Wallet " ++ show n ++ " mainnet address: " ++ mainnetAddress w
 
     toAda = (* 1_000_000)
-
-    -- waits for the last wallet to be funded
-    awaitFunds ws delay = do
-      let lastWallet = last ws
-      liftIO $ putStrLn "Waiting till all wallets will be funded..."
-      awaitAddressFunded (cardanoMainnetAddress lastWallet) delay
-
-termHandler :: TVar (ClusterStatus ()) -> System.Posix.Handler
-termHandler st = CatchOnce $ do
-  putStrLn "Caught SIGTERM, stopping cluster"
-  stopCluster st
 
 data ClusterInfo = ClusterInfo
   { ciWallets :: [(String, String)]
@@ -219,26 +188,6 @@ pEpochSize =
             <> Options.value 160
         )
 
-pChainIndexMode :: Parser ChainIndexMode
-pChainIndexMode =
-  noIndex <|> withIndexPort <|> pure DefaultPort
-  where
-    noIndex =
-      Options.flag'
-        NotNeeded
-        ( Options.long "no-index"
-            <> Options.help "Start cluster with chain-index on default port"
-        )
-    withIndexPort = CustomPort <$> portParser
-
-    portParser =
-      Options.option
-        Options.auto
-        ( Options.long "chain-index-port"
-            <> Options.metavar "PORT"
-            <> Options.help "Start cluster with chain-index on custom port"
-        )
-
 pInfoJson :: Parser (Maybe FilePath)
 pInfoJson =
   optional $
@@ -260,7 +209,6 @@ pClusterConfig =
     <*> pWorkDir
     <*> pSlotLen
     <*> pEpochSize
-    <*> pChainIndexMode
     <*> pInfoJson
 
 -- | Basic info about the cluster, to
@@ -274,7 +222,6 @@ data ClusterConfig = ClusterConfig
   , workDir :: Maybe FilePath
   , slotLength :: NominalDiffTime
   , epochSize :: EpochSize
-  , cIndexMode :: ChainIndexMode
   , dumpInfo :: Maybe FilePath
   }
   deriving stock (Show, Eq)
