--- src/Test/Plutip/Internal/BotPlutusInterface/Types.hs
@@ -1,8 +1,0 @@
-module Test.Plutip.Internal.BotPlutusInterface.Types (
-  BpiError (..),
-) where
-
-data BpiError
-  = SignKeySaveError !String
-  | BotInterfaceDirMissing
-  deriving stock (Show)
