+++ test/Main.hs
@@ -1,8 +1,7 @@
 module Main (main) where
 
-import Spec.Integration qualified as Integration
+import Spec.ClusterStartup qualified as ClusterStartup
 
-import Spec.Test.Plutip.BotPlutusInterface qualified as BotInterface
 import Test.Tasty (defaultMain, testGroup)
 
 main :: IO ()
@@ -10,8 +9,5 @@ main =
   defaultMain $
     testGroup
       "tests"
-      -- FIXME: both `Integration.test` and `BotInterface.test`
-      -- start own cluster to run tests, probably, need better solution in future
-      [ Integration.test
-      , BotInterface.test
+      [ ClusterStartup.test
       ]
